api/requirements.txt
django
djangorestframework
django-widget-tweaks
django-webpack-loader 
gunicorn
psycopg2-binary
python-dotenv
netCDF4
django-cors-headers
django-extensions
scipy
selenium
